overhaul

+++ contracts/game/Irng6.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.6.12;
 
 interface IRandomNumberGenerator {
-    function getRandomNumber() external view returns (uint randomness);
+    function getRandomNumber() external view returns (uint256 randomness);
 }
